move block helpers to separate file
