+++ src/mindustry/block/crafting.ts
@@ -1,4 +1,5 @@
-import { Block, BlockOutput, BlockOutputDirection } from './block'
+import { BlockOutput, BlockOutputDirection } from './helper'
+import { Block } from './block'
 import { ItemCost } from '../item/item_cost'
 import { RenderingInfo } from '../../util'
 import { SchematicTile } from '../../schematic'
