+++ src/mindustry/block/experimental.ts
@@ -1,4 +1,5 @@
-import { Block, BlockOutput, BlockOutputDirection } from './block'
+import { BlockOutput, BlockOutputDirection } from './helper'
+import { Block } from './block'
 import { RenderingInfo } from '../../util'
 import { SchematicTile } from '../../schematic'
