+++ src/mindustry/block/index.ts
@@ -1,2 +1,3 @@
-export { Block, BlockOutput } from './block'
+export { Block } from './block'
 export { Blocks } from './blocks'
+export * from './helper'
