+++ src/mindustry/block/storage.ts
@@ -1,5 +1,6 @@
-import { Block, BlockOutput, BlockOutputDirection } from './block'
+import { BlockOutput, BlockOutputDirection } from './helper'
 import { RenderingInfo, blockAsset, tintImage } from '../../util'
+import { Block } from './block'
 import { Item } from '../item'
 import { SchematicTile } from '../../schematic'
