+++ src/mindustry/block/units.ts
@@ -1,4 +1,4 @@
-import { Block, BlockOutput, BlockOutputDirection } from './block'
+import { BlockOutput, BlockOutputDirection } from './helper'
 import {
   RenderingInfo,
   blockAsset,
@@ -6,6 +6,7 @@ import {
   outlineImage,
   tintImage,
 } from '../../util'
+import { Block } from './block'
 import { ItemCost } from '../item'
 import { SchematicTile } from '../../schematic'
